// src/lib/schemas/actor.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Maree Backend API
 * OpenAPI spec version: 1.0.0
 */

export interface Actor {
  id: string;
  firstName: string;
  lastName: string;
  phone: string;
  /** @nullable */
  email: string | null;
  createdAt: string;
  /** @nullable */
  role: string | null;
}